Gestion de fichiers, post sharp

+++ backend/controllers/books.js
@@ -1,7 +1,8 @@
-const Book = require('../models/Book')
+const Book = require('../models/Book');
+const fs = require('fs');
 
 exports.addBook = (req, res, next) => {
-    const bookObject = JSON.parse(req.body.thing);
+    const bookObject = JSON.parse(req.body.book);
     delete bookObject._id;
     delete bookObject._userId;
     const book = new Book({
@@ -15,20 +16,47 @@ exports.addBook = (req, res, next) => {
     .catch(error => { res.status(400).json( { error })})
 };
 
-// Modification de l'objet
+// Modification de l'objet // On gère ici deux cas : Si l'utilisateur a transmis un fichier ou non lors de la modification
 exports.modifyBook = (req, res, next) => {
-    Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id}) // Le 1er argument de updateOne est l'objet de comparaison (celui qu'on va modifier)
-                                                                                // Le 2ème argument est le nouvel objet. On récupère le Book dans le corps de la requête avec
-                                                                                // '...req.body', et on précise que l'id correspond à celui des paramètres.
-    .then(() => res.status(200).json({ message: 'Objet modifié'})) // Code 200 = requête client faite avec succès
-    .catch((error => res.status(400).json({ error})));
+    const bookObject = req.file ? { // On regarde s'il y a un champ 'file' dans l'objet requête
+        ...JSON.parse(req.body.book), // Si oui, on parse la chaine de caractères dans un json et on crée une URL au fichier image
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body }; // Si pas de 'file' on récupère directement l'objet dans le corps de la requête
+
+    delete bookObject._userId;
+    Book.findOne({_id: req.params.id}) // On récupère l'objet dans la BDD pour vérifier que l'utilisateur qui cherche à le modifier soit bien celui qui l'a créé
+        .then((book) => {
+            if (book.userId != req.auth.userId) { // Si le userId de la BDD est différent de celui qui vient du token, ça veut dire que qqun tente de modifier un objet qui ne lui appartient pas
+                res.status(401).json({message: "Non-autorisé" });
+            } else { //  ({enregistrement à mettre à jour}, {avec quel objet}
+                Book.updateOne({_id: req.params.id}, {...bookObject, _id: req.params.id})
+                .then(() => res.status(200).json({message: "Objet modifié"}))
+                .catch(error => res.status(401).json({ error }));
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error })
+        })
 };
 
-//Suppression d'un objet
+//Suppression d'un objet 
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id }) // On sélectionne l'objet par son id en vérifiant qu'il correspond à l'id des paramètres de requête
-    .then(() => res.status(200).json({message: 'Livre supprimé'}))
-    .catch(error => res.status(400).json({ error}));
+    Book.findOne({_id: req.params.id}) // On récupère l'objet dans la BDD
+    .then(book => {
+        if (book.userId != req.auth.userId){ // On vérifie que le userId enregistré en BDD correspond à celui qui vient du token
+            res.status(401).json({message: "Non-autorisé"});
+        } else {
+            const filename = book.imageUrl.split('/images/')[1]; // On récupère le nom du fichier grâce à l'URL de l'image. On fait un split pour récupérer l'élément qui suit '/images/' ( [1] )
+            fs.unlink(`images/${filename}`, () => { 
+                Book.deleteOne({_id: req.params.id}) // On cible l'Id pour sélectionner le livre à supprimer
+                .then(() => {res.status(200).json({message: 'Livre supprimé'})})
+                .catch(error => res.status(401).json({ error }));
+            });
+        }
+    })
+    .catch( error => {
+        res.status(500).json({ error });
+    })
 };
 
 // Ceci est la requête GET qui va afficher les objets en fonction de l'ID récupérée (:id), elle les sélectionne donc de façon individuelle
